// JavaScript/PokemonSearch/src/Types/CheckingShinyPokemonNullData.tsx
// CheckingShinyPokemonNullData.tsx

export type CheckingShinyPokemonNullData = {
    [key: string]: string | undefined;
        redblue?: string;
        yellow?: string;
        gold?: string;
        silver?: string;
        crystal?: string;
        rubysapphire?: string;
        fRedlGreen?: string;
        diamondpearl?: string;
        platinum?: string;
        heartgoldsoulsilver?: string;
        blackwhite?: string;
        blackwhiteAnime?: string;
        default?: string;
}